requirements.txt
# Chandigarh Policy Assistant Requirements - Railway Deployment
# Optimized for Jina v3 API (No Local Models Required)

# Core Web Framework
flask>=2.3.0,<3.0.0
flask-cors>=4.0.0,<5.0.0

# Vector Database & AI APIs  
pinecone[grpc]>=4.0.0,<5.0.0
groq>=0.4.0,<1.0.0

# Essential ML/Data Processing (Lightweight)
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# Search and NLP (Lightweight)
rank-bm25>=0.2.2,<1.0.0
nltk>=3.8,<4.0.0

# HTTP and Utilities
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Optional utilities (lightweight)
tabulate>=0.9.0,<1.0.0
tqdm>=4.27,<5.0.0

# Railway deployment optimization
gunicorn>=21.0.0,<22.0.0 